Move framework defaults into core to break the core->implementations cycle
pySDC/core/controller.py imported CheckConvergence, DefaultHooks and
CPUTimings from pySDC/implementations/, an upward dependency from the
abstraction layer into the implementation layer. That made the kernel
depend on its own plug-ins and put a cycle in the package graph.

These three are not optional plug-ins - the framework cannot run without
them - so they are reclassified as core rather than made injectable:

  implementations/convergence_controller_classes/check_convergence.py
    -> core/check_convergence.py
  implementations/hooks/default_hook.py -> core/default_hook.py
  implementations/hooks/log_timings.py  -> core/log_timings.py

log_timings.py is split: Timings and CPUTimings move to core, while
GPUTimings stays a plug-in in implementations/hooks/ since it needs the
optional cupy dependency. Importing pySDC.core.* now pulls in no
implementations module at all.

BREAKING: the old import paths are gone, with no re-export shim. All 16
in-repo call sites are updated. Consistent with how Step already handles
the removed dtype_u/dtype_f parameters - a hard error rather than a shim.
GPUTimings' import path is unchanged.
