+++ pySDC/core/check_convergence.py
