+++ pySDC/projects/deprecated/node_failure/controller_nonMPI_hard_faults.py
@@ -1,5 +1,5 @@
 from pySDC.implementations.controller_classes.controller_nonMPI import controller_nonMPI
-from pySDC.implementations.convergence_controller_classes.check_convergence import CheckConvergence
+from pySDC.core.check_convergence import CheckConvergence
 
 from pySDC.projects.node_failure.emulate_hard_faults import hard_fault_injection
